refactor(cli): put ops in their own modules
