+++ crates/mun/src/ops.rs
@@ -1,0 +1,11 @@
+mod build;
+pub mod init;
+mod language_server;
+mod new;
+mod start;
+
+pub use build::build;
+pub use init::init;
+pub use language_server::language_server;
+pub use new::new;
+pub use start::start;
